+++ CHANGELOG.md
@@ -55,3 +55,9 @@
 # v1.2.0 - Released on 2019/12/16
 
 - Add: export ESM and full-featured file
+
+# v1.3.0 -
+
+- Add: `slidesPerView` parameter
+- Add: `centeredSlides` parameter
+- Fix: Unexpected slide width (#14)